Kiwi cam mergetest (#69)
* Support various new RM Mini devices by updating to the latest broadlinkjs-rm library
* Support for resetting to a default fan speed each time a fan is turned on
          "alwaysResetToDefaults": true,
          "defaultFanSpeed": 30,
* Support for a “window” accessory that mimic the functionality of a window-covering
* Added support for Air Purifier accessories
* Added default state for lock controls
* Added support for Humidifier/Dehumidifiers 
* Garage Door Opener - If the `autoCloseDelay` is set to `0`, set the Homekit state to closed immediately
* Allow for separate commands for turning swing mode/oscillation on/off 
* Fixed a potential crash when learning RF data
* Updated packages to resolve potential security issues.
* Improvements to how ping’s work to improve performance and device re-discovery

+++ platform.js
@@ -10,6 +10,8 @@ const { createAccessory } = require('./helpers/accessoryCreator');
 
 const classTypes = {
   'air-conditioner': Accessory.AirCon,
+  'air-purifier': Accessory.AirPurifier,
+  'humidifier-dehumidifier': Accessory.HumidifierDehumidifier,
   'learn-ir': Accessory.LearnCode,
   'learn-code': Accessory.LearnCode,
   'switch': Accessory.Switch,
@@ -21,6 +23,7 @@ const classTypes = {
   'fan': Accessory.Fan,
   'outlet': Accessory.Outlet,
   'light': Accessory.Light,
+  'window': Accessory.Window,
   'window-covering': Accessory.WindowCovering,
   'tv': Accessory.TV
 }
@@ -61,6 +64,9 @@ const BroadlinkRMPlatform = class extends HomebridgePlatform {
     // Iterate through the config accessories
     let tvs = [];
     config.accessories.forEach((accessory) => {
+      // Optionally hide the accessory from Homekit
+      if (accessory.ignore === true) return
+
       if (!accessory.type) throw new Error(`Each accessory must be configured with a "type". e.g. "switch"`);
 
       if (!classTypes[accessory.type]) throw new Error(`homebridge-broadlink-rm doesn't support accessories of type "${accessory.type}".`);
